Fix a couple of bugs in subidAlloc package.
This change fixes a couple of problems in the subidAlloc package,
introduced by 4e5e8d9357e1c836bd2b6379be6ab7b3149d39e7.

That commit broke all of the unit tests, and also had a couple of
bugs in the alloc and free implementations.

This change fixes all of the above.