First commit of shiftfs mount manager in sysbox-mgr.
This is required for the migration of sysbox from using
the Nestybox shiftfs module to the Ubuntu shiftfs module.

The shiftfs mount manager in the sysbox-mgr does
shiftfs marks on each sys container's rootfs or mount sources
when the container starts, and removes them when the container
stops.

If multiple sys containers share a mount source, it ensures that
shiftfs marking is done only once and removes the marking only
when no sys containers are using the mount point.