Initial implementation of sysbox-mgr subid allocator identity-map mode.
This feature causes the sysbox-mgr subid allocator to map uid(gid)
range 0->65536 in the system container to 0->65536 on the host.

The goal of this mode is to enable Docker to spawn system containers
when configured with userns-remap disabled, and without requiring
Sysbox to use uid-shifting (i.e., shiftfs). This is desirable because
when using uid-shifting some functional restrictions currently apply
(e.g., sysbox issue #336).

Note however that this feature reduces isolation between the system
contaiener and the host because root in the system container is root
on the host (though the user-namespace provides protection because
a system container root process has no capabilities outside of the
system container's user namespace).

Bonus: minor cleanup of sysbox-mgr subid allocator knobs.