Ensure subid allocator returns same uid and gid for each allocation.
The sysbox-mgr subid allocator had a bug that was causing it
to sometimes allocate different uid and gid ranges for a given
system container.

This was causing the system container to fail when using uid-shifting,
since shiftfs requires that a given mount have the same shifting for
both uid and gid.

This bug was causing the failure reported in sysbox issue #229.

This change fixes this by ensuring the sysbox-mgr subid allocator
always returns the same uid and gid for a given system container.