Add support for modifying uid(gid) of mount sources mounted at the sys container's /var/lib/docker.
This change modifies sysbox-mgr to modify the uid(gid) of mount sources
mounted at a sys container's /var/lib/docker.

The ownership change is done in order to ensure the sys container
sees "root:root" ownership on the contents of the mount. It avoids
using shiftfs or userns identity-maps on the mount (shiftfs is
not an option as overlayfs can't be mounted on top of it, and
identity-mapping is not a good choice from a security perspective).

The ownership change is only done when sysbox operates in
exclusive userns-remap mode.