Add inner Docker image sharing among system containers.
This change introduces the dockerVolMgr, which is a package that
managers host volumes that are bind-mounted into the sys container's
/var/lib/docker.

Unlike the regular volMgr package, the dockerVolMgr is capable of
sharing inner Docker image layers across multiple sys containers.
That is, when multiple sys containers associated with the same
Docker image are launched, and said Docker image has inner Docker
images within it, the dockerVolMgr allows those sys containers
to share the inner Docker images using a copy-on-write mechanism.

This has two key benefits:

1) It reduces storage usage for sys containers that come preloaded
with large inner images.

2) It reduces sys container startup time (when multiple sys containers
sharing the same image are launched, the startup time of all but the
first container is reduced).

These benefits are larger for sys container images that come preloaded
with large number or heavy inner container images (e.g., such as a
K8s node sys container image).

For example, a sys container k8s node image can be as large as
1.7GB. With the regular volMgr, starting 10 sys containers from this
image would eat up (1.7 x 10) = 17GB of storage (!), and take 10
seconds each (just to start the sys container!).

With the dockerVolMgr, it eats up 1.8GB total, and the first container
takes ~ 6 secs to start, and the remained take ~ 2 secs to start.

This change includes unit and integration tests for the dockerVolMgr.