Added preflight check to catch sysbox-mgr host requirements upon start.
This is helpful as I found that sysbox-mgr was failing on my host because
I had inadvertendly removed rsync (which sysbox-mgr depends on).

Since rsync was missing, sysbox-mgr failed to unregister a sys
container (because it could not perform a copy during that time). The
failure was not obvious: even though sysbox-mgr returned an error to
sysbox-runc, the higher level tool I was using to run the containers
(k8s.io KinD) did not report the error.

With this change, we catch depedency problems upfront. It's not a
perfect check because a user can always remove a program after
installing sysbox-runc, but it's helpful.