Create a sysbox-mgr volume manager to back a sys container's /var/lib/containerd dir.
In sys containers that use containerd to pull inner container images,
these images are unpacked into a sub-dir of /var/lib/containerd.

Turns out that the unpacking operation fails if shiftfs is mounted
on /var/lib/containerd (see Sysbox issue #622 for details). To work-around
this, this commit causes the Sysbox-mgr to back the sys container's
/var/lib/containerd directory with a directory on the host. This way,
the mount of shiftfs is avoided.