Move linux header and module mounts setup to sysbox-mgr.
This setup used to be done in sysbox-runc, but is being moved
to sysbox-mgr in order to reduce sysbox-runc container creation
latency. Sysbox-mgr can pre-compute the linux header and module
mounts setup for the sys containers at startup time, thus
avoiding the need to perform the computation whenever a container
is created.