Remove check for sysbox-mgr pid file during start.
The sysbox-mgr pid file was added as a way of ensuring
that only one sysbox-mgr instance runs at a time.

But this check is really not required, since sysbox is designed
to work with systemd which will ensure this.

Plus the sysbox-mgr pid file ended up causing problems in
situations where sysbox-mgr crashed and did not remove the file,
thereby preventing a new instances of sysbox-mgr from starting.