Make unregister quicker by skipping sync-out on Docker containers launched with "--rm".
This change enables sysbox-mgr to detect containers that were launched with
"--rm". For such containers, sysbox-mgr can skip copying the contents
for the docker, containerd, and kubelet volumes back to the container's
rootfs. That's because there is no point in doing this copy if we know
that the container will be removed anyway.

This in turn makes the docker stop for a sys container started with "--rm"
faster, in particular for a sys container that stores data in
its /var/lib/containerd, /var/lib/docker, or /var/lib/kubelet
directories.