Fix for issue preventing linux-headers from being properly exposed within non-ubuntu containers
Within the context of the linux-headers' exposure problem inside a sys-container, Sysbox-mgr has till now assumed that the linux distribution at the host and the sys-container level will always match. In heterogeneous environments, a sys-container running a distro different from the host one, may expect kernel-headers in a different path than the host system. We fix this problem as explained below:

1) We continue to honor the host's kernel-header path by bind-mounting this host's header folder in the same path within the sys-container (regardless of the distro running in it).

2) To satisfy the per-distro expected path, we create a soft-link inside the sys-container that points to the real header that bind-mounted in step 1).

Note: To identify the kernel-header path expected by each sys-container's distro we rely on this classification:

- centos / redhat / fedora: "/usr/src/kernels/<blah>"

- ubuntu / debian / others: "/usr/src/linux-headers-<blah>-generics"

This distro-to-path logic will need further refinement as we include support for more distros.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>