Fix for issue preventing linux-headers from being properly exposed within non-ubuntu containers
Sysbox-mgr was previously hard-coding the location where linux-headers (in the host system) would be bind-mounted into the sys-container. The previous logic was assuming that other linux-distros would also expect linux-headers in the same location (i.e. /usr/src/linux-headers-<*>-generic).

During sys-container initialization now we identify the linux-distro within the container's rootfs, and based on that, we bind-mount the host's linux-headers into the path expected by that specific linux-distro.

Notice that we are relying on the two most common patterns seen thus far:

- centos / redhat / fedora: "/usr/src/kernels/<blah>"

- ubuntu / debian / others: "/usr/src/linux-headers-<blah>-generics"

This distro-to-path logic will need further refinement as we include support for more distros.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>