Ensure that sysbox-mgr connection to Docker API is properly closed.
Prior to this change, sysbox-mgr had a bug where it was not properly
closing the file descriptor handle for the connection to the
Docker API. As a result, the sysbox-mgr was leaking file-descriptors,
which would cause failures after hitting the Linux limit for
max file descriptors opened by a process (typically 1024).

This change fixes this by ensuring sysbox-mgr closes the
Docker API connection file descriptor appropriately.

Signed-off-by: ctalledo <ctalledo@nestybox.com>