Fix bug parsing of subid ranges in /etc/login.defs.
When parsing the file, entries may be separated by any unicode space characters
(e.g., space, tab, etc.). Prior to this change, the code was assuming it had to
be the space character. This failed when running Sysbox in a Fedora-32 kernel,
in which tabs where used to separate the fields.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>