Change prepMounts() API to always chown mount sources.
Bind-mounts over special container dirs such as "/var/lib/docker",
"/var/lib/kubelet", etc., always need Sysbox to perform a
chown to ensure the mount source matches the container's uid:gid.

This change ensures this. Prior to this change, the chown was
conditionalized on whether the container's rootfs needed
uid shifting, but this is not correct.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>