Relax error check during container unregistration.
This change relaxes an error check during container unregistration. It's needed
to deal properly with cases where a container registers and then unregisters due
to an error that occurs when starting the container (e.g., on sysbox-runc).

Without this change, such a situation results in sysbox-mgr generating an
unnecessary warning on its log.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>