Fix: longestCommonPath should not truncate a path component
Before this commit longestCommonPath did not contain any logic to deal with
path components, so some input like {"/usr/src/abc", "/usr/src/abd"} returned
"/usr/src/ab" instead of "/usr/src/".

This commit adds the required logic to detect when we truncated at the middle
or a path component and if so, strip this last path component.

As far as I know, this bug did not have any impact on any supported
configuration, it was accidentally found during manual testing.

Signed-off-by: Joan Bruguera <joanbrugueram@gmail.com>