Perform proper uid shifting via chown on special container mounts.
For containers where the user wants to mount a host dir on a container directory
which Sysbox considers "special" (e.g., /var/lib/docker, /var/lib/kubelet,
etc.), Sysbox needs to do some uid-shifting on the host directory using chown
(i.e., shiftfs can't be mounted on these special dirs).

Prior to this change, the chown operation was not doing proper shifting of the
host directory user and group IDs, but rather changing them from root on the
host to root in the container when the container started, and vice-versa when
the conatiner stopped. This is tracked by Sysbox issue #251.

This change fixes this, though it's still missing dealing with
Linux access control lists (ACL) when chown'ing files.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>