Fix performance regression caused by prior change on uid shifting.
A recent change in the way sysbox-mgr performs uid-shifting
of so called "special dirs" in the container (e.g., /var/lib/docker)
caused a severe performance regression which is noticeable
when stopping a sysbox container that comes preloaded with
lots of data in those special dirs (e.g., a sysbox container
that comes preloaded with heavy inner docker images).

The performance regression was due to the fact that during
container creation, sysbox-mgr copied the files in the
container's special dir to a sysbox managed dir using
rsync and then modified the ownership. This causes
rsync to think all fines have changed. As a result,
when the container stops, the copy-back operation
(from the sysbox managed dir to the container's rootfs)
becomes very slow (rsync copies all files).

This commit fixes this by having rsync do the chown
during the copy. It does this using rsync's --usermap
and --groupmap options. This way, when the container
stops, rsycn will only copy back files that
changed from the sysbox managed dir to the container's
rootfs, improving performance significantly
with respect to the prior commit that caused the
peformance regression.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>