Improve uid-shifting of ACLs to support files that can't be opened.
When Sysbox does uid-shifting of host directories mounted into the Sysbox
container's special dirs (e.g., host directories mounted to the container's
/var/lib/docker), some of those files may be special character files
that can't be opened (e.g., overlayfs whiteouts for example).

Prior to this commit Sysbox was opening the file before performing
the uid shift on the ACLs. But opening the file is not necessary
as we only need to call the setxattr() and getxattr() syscalls
to perform the operation.

This commit fixes this.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>