Changes to improve security of shiftfs mounts.
This commit modifies the way in which Sysbox sets up shiftfs mounts for host
files / directories bind mounted into a container.

Prior to this change, Sysbox applied the shiftfs mount directly on the directory
that was bind-mounted to the container (or if a file was bind-mounted, on the
file's directory).

But this is not ideal because it means that in hosts where regular users are
allowed to create user-namespaces, a regular user could enter a user-ns and
mount shiftfs on these same host directories, thereby gaining root access to
them.

This commit mitigates this by having Sysbox setup the shiftfs mounts in a
directory that is only accessible by root (e.g., /var/lib/sysbox/shiftfs),
thereby preventing regular users from mounting shiftfs on them.

For example, if a container has a bind-mount of host dir /home/user/somedir,
Sysbox will no longer set the shiftfs mount directly on /home/user/somedir.
Rather it will mount /home/user/somedir under /var/lib/sysbox/shiftfs/<uuid>,
mount shiftfs on the latter, and bind-mount it into the container.

This way, a regular user can never gain access to the shiftfs mount
since it's a root-only accessible directory (/var/lib/sysbox/shiftfs).

Many thanks to Joan Brugera (@joanbm) for catching and reporting this.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>