Fix Linux header mounts when restarting a container
When restarting (docker stop + docker start) a container, the Linux
kernel headers were not mounted again. The reason is quite simple,
which is that reqMounts in sysbox-mgr attempts to cache the mounts,
but it was returning a list which only contains some of them (special
system container mounts only) but not the Linux kernel headers.

Store the full list of mounts and use it when restarting a container.

Manual test case:
    docker run --runtime=sysbox-runc -d --name testcnt \
        alpine:3.13 sleep 999999
    docker exec testcnt mount # Kernel headers OK
    docker stop testcnt
    docker start testcnt
    docker exec testcnt mount # Kernel headers missing

Signed-off-by: Joan Bruguera <joanbrugueram@gmail.com>