Capture special devices during inner Docker image preloading.
Sysbox supports preloading of inner Docker images into a Sysbox
container, via a "docker build" or "docker commit".

However, there was a problem when the inner images being preloaded had
special devices in them (e.g., FIFOs and sockets). This problem led
to the preloaded images missing the special devices, causing problems
for applications that depended on these.

This commit fixes this.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>