Modprobe configfs when starting Sysbox.
This is required so that mounting procfs inside sys containers creates a
configfs folder, on top of which Sysbox currently creates a dummy mount.

Typically the modprobe is done by the Sysbox package installer, but for
scenarios where Sysbox was installed in a different way (e.g., built from
source), the modprobe during Sysbox startup helps ensure configfs
is present.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>