Deprecate bind-mount-id-shift config; replace with disable-shiftfs and disable-idmapped-mount options.
This change replaces the bind-mount-id-shift config (never used) and replaces it
with the disable-shiftfs and disable-idmapped-mount options.

The latter options will be helpful for testing purposes, as they will allow the
Sysbox test suite to test behavior with/without shiftfs/ID-mapped-mounts on a
single host machine that supports both.

Note that when both shiftfs and ID-mapped-mounts are enabled, Sysbox uses
ID-mapped-mounts. When both are disabled, Sysbox uses chown on the container's
rootfs but not on volume mounts, causing the latter to show up as nobody:nogroup
inside the container. See the sysbox-runc code for more info.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>