First commit of rootfs clone manager.
This manager handles cloning of a container's rootfs, which is done whenever
Sysbox needs to chown the container's rootfs because the kernel does not support
uid shifting on overlayfs mounts (i.e., no shiftfs).

If the rootfs is mounted on overlayfs without metacopy=on (e.g., as done by
Docker), the the chown can be very slow (several seconds). Moreover, enabling
metacopy=on on the mount breaks container snapshot operations operations like
"docker commit" since they capture file metadata instead of full files.

To solve this, the rootfs clone manager clones the container's original rootfs
and uses a couple a technique to ensure the chown can go fast yet does not
break container snapshots.

This is accomplished by cloning the rootfs using two stacked overlayfs
mounts, called "top" and "bottom" mounts.

The bottom mount is identical to the overlayfs mount of the original
rootfs, except it adds "metacopy=on" to enable fast chowns on it.
This enables very fast chown of the container's rootfs.

The top mount uses the bottom mount's "merged" dir as its lower
layer. The merged dir of the top mount is the container's rootfs,
while the upper dir of the top mount tracks the rootfs diffs generated
by the container. Since the top mount does not use metacopy=on,
container snapshot operations work properly.

The only problem is that this pollutes the host mount-ns with
more mounts, but it's the best solution we found at this time.

Note that the need for the rootfs cloner will go away once
the Linux kernel supports idmapped mouts on top of overlayfs.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>