Add ignore-sysfs-chown config option to sysbox-mgr.
The /sys inside Sysbox containers shows up as nobody:nogroup, because it's owned
by the host root.

Some apps inside the container (e.g., RedHat package manager (rpm)) some times
try to chown /sys to root:root. This operation fails with EPERM, causing the app
to fail.

To work-aroundthis, Sysbox currently ignores chown of /sys inside the container.
However, this requires that Sysbox trap the chown syscall inside the container,
which has a negative effect on performance in several cases.

With this commit, we modify the default behavior of Sysbox to not allow
chown of /sys inside the container, but provide a config option to allow
it. This way, most users won't pay the performance penalty of having
Sysbox trap chown, yet for the few use cases where it's needed, Sysbox
offers a solution.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>