Add --syscont-mode command line flag to sysbox-mgr.
When set to true (default), this global flag causes Sysbox to set up the
container as a "system container" (i.e., to run system workloads on it).

When set to false, this flag causes Sysbox to set up the container as a "regular
container" (typically to run microservices), by closely following the
container's OCI spec, except for the container's user-ns and cgroup-ns, which
Sysbox always enables on all containers for extra security.

The --syscont-mode flag is a global flag: when set to true or false, it applies
to all containers deployed with Sysbox. However, the flag can be overriden on a
per-container basis via the SYSBOX_SYSCONT_MODE=[TRUE|FALSE] env var passed to
each container.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>