Have sysbox-mgr check if ID-mapping is supported on host.
This commit makes sysbox-mgr check if ID-mapping is supported
on the host. The check is performed during initialization.

sysbox-mgr will check if ID-maping is supported (kernel >= 5.12)
and if ID-mapping on overlayfs is supported (kernel >= 5.19).
For the latter, the check is done empirically by running a
quick test on the host.

The info on ID-mapping support is passed by sysbox-mgr to sysbox-runc when a
container is registered. sysbox-runc will use this info to set up the container.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>