Chown rootfs overlayfs upper dir when ID-mapped.
When the container's rootfs is on overlayfs, and Sysbox is using
ID-mapped-mounts on it, then the ID-mapping only applies to the
overlayfs lower layers, but not the upper layer.

As a result, when the container stops or pauses, Sysbox must do a chown of the
rootfs upper layer from container host ID range -> [0->64k] (e.g., from uid:gid
165536:165536 -> 0:0) so that it matches the ownership of the lower layers
(which remain in the range [0->64K]).

If Sysbox does not do this, then a container commit won't work properly
when the container restarts.

This commit does this chown of the rootfs ovelayfs upper dir
when necessary.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>