Add --disable-inner-image-preload flag to sysbox-mgr.
This flag disables the Sysbox feature that allows for preloading of inner
container images into a Sysbox container (e.g., via Docker commit or build),
or for running containers that come preloaded with inner images.

While the image preloading feature is useful in some cases, it increases the
amount of time to start or stop a container. The increase can be significant in
scenarios where Docker or K8s runs inside the Sysbox container and pulls many
inner images (or heavy inner images).

By setting --disable-inner-image-preload, users can avoid slowing down container
start/stop time, at the cost of loosing the feature that allows for inner Docker
image preloading. This trade-off is useful for many users, particularly since
most users don't use inner image preloading.

Note: in the future we may offer a way to set this per-container (as opposed to
globally via the --disable-inner-image-preload flag).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>