Simplify rootfs cloning to prevent inode leakage.
NOTE: This commit fixes the inode leakage reported in Sysbox issue 570
but has the undesired effect that "docker build" or "docker commits" with sysbox
as the default runtime will not work on hosts with kernel < 5.19 and without
shiftfs. If the host has kernel 5.19+ or uses shiftfs, then everything works
as usual.

Description:

Per Sysbox issue 570 (https://github.com/nestybox/sysbox/issues/570), creation
of sysbox containers on hosts with kernel < 5.19 and without shiftfs results in
inode leakage.

That is, when the container is created, inodes are consumed; when the container
is later destroyed, those inodes are not returned back to host.

The problem appears to be a kernel or overlayfs issue triggered by Sysbox's
rootfs cloning code in sysbox-mgr. The rootfs cloning code was creating two
stacked overlayfs mounts, where the "merged" dir of the bottom mount was used as
a lower layer of the top mount. This served two purposes:

* Bottom mount: it was a clone of the container's rootfs overlay mount, but with
  metacopy=on. This allowed sysbox to chown the container's rootfs very quickly
  on container startup and revert the chown on stop.

* Top mount: it used the bottom mount's "merged" dir as a lower layer, and acted
  as the container's actual rootfs. This way the container would see the effect
  of the chown in the bottom mount. In addition, the top mount has metacopy=off,
  such that its "diff" dir would track changes to the container's filesystem at
  runtime, thereby making container image commit operations (e.g., docker build
  and docker commit) work well.

Unfortunately, the overlayfs stacking causes inode leakage when the container
stops.  It's not clear why, but it appears that the kernel or overlayfs is not
releasing the inodes for the bottom mount when the container stops, even though
no processes exist in the container and sysbox tears down the top and bottom
mounts (i.e., no processes are using the stacked overlayfs mounts). The inodes
associated with the bottom mount become "orphaned inodes", which consume storage
and only get cleaned up upon the host's reboot.

To work-around the inode leakage problem, this commit gets rid of the overlayfs
mount stacking, and keeps the bottom mount only. This fixes the inode leakage,
but has the UNDESIRED EFFECT THAT "DOCKER BUILD" OR "DOCKER COMMITS" WITH SYSBOX
AS THE DEFAULT RUNTIME WILL NOT WORK on hosts with kernel < 5.19 and without
shiftfs.

Though this is not ideal, it's certainly more important to close the inode
leakage as that problem occurs when running Sysbox containers and can eventually
result in "out of space" issues on the host.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>