Fix bug in shiftfs and ID-mapping precheck.
When sysbox-mgr starts, it checks if shiftfs and ID-mapping work on the host. It
does this by creating spawning a child process into a Linux user-namespace, and
having the child create a temporary dir under /var/lib/sysbox, and then use it
to mount shiftfs or ID-mapped mounts.

The bug is that /var/lib/sysbox is setup by Sysbox with 0710 permissions, and
this prevents the child process from performing the test correctly in some
hosts.  The fix is to modify the permissions to 0755 temporarily during the
shiftfs/ID-mapping precheck, and then back to 0710 once the check completes.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>