Reduce effect of disable-inner-image-preload flag.
Prior to this change, the disable-inner-image-preload flag was not only
disabling the preloading of inner container images into system containers, but
also running of system containers that come with preloaded images.

The latter creates an incompatibility issue as it causes Sysbox to be unable to
run existing system containers that come with preloaded inner images.

This commit removes this incompatibility. That is, when
"--disable-inner-image-preload" is set to true, a user won't be able to use
"docker commit" or "docker build" to preload inner container images into a
Sysbox container. However, the user will continue to be able to run existing
Sysbox container images that come preloaded with inner containers (e.g.,
nestybox/k8s-node images).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>