Fix rootfs ID-mapping bug exposed by docker -w.
NOTE: for context, refer to the corresponding commit in sysbox-runc.

In this commit, we adjust sysbox-mgr to track the fact that sysbox-runc
may chown the container's rootfs overlay upper layer. This way,
sysbox-mgr can revert that chown during container stop or pause,
and re-instate it during container resume (sysbox-runc only
does the chown during container create).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>