More fixes in host kernel header mounting.
When parsing the list of host kernel headers to be mounted into Sysbox
containers (typically under /usr/src), the sysbox-mgr finds the headers in the
host and follows any symlinks, and then tries to determine the lowest common
path (lcp) for bind-mounting the headers (so as to avoid creating too many
bind-mounts into the Sysbox container).

The computation of the lcp was not handling properly the scenario
when the kernel headers are located in different directories under
/usr/src/.

For example, in the scenario where a host has this kernel header config:

/usr/src/linux-headers-6.2.0-35-generic/block -> ../linux-hwe-6.2-headers-6.2.0-35/block
/usr/src/linux-headers-6.2.0-35-generic/certs -> ../linux-hwe-6.2-headers-6.2.0-35/certs
...
/usr/src/linux-headers-6.2.0-35-generic/rust -> ../linux-hwe-6.2-lib-rust-6.2.0-35-generic/rust

the sysbox-mgr was computing the lcp as "/usr/src" and trying to bind-mount that
directory into the Sysbox container. That's because when following the symlinks,
the "*rust*" headers where not under "/usr/src/linux-hwe-6.2-headers-6.2.0-35"
but rather under "/usr/src/linux-hwe-6.2-lib-rust-6.2.0-35-generic/rust", so the
lowest common path is "/usr/src/".

But this is not ideal since "/usr/src" contains other stuff that we don't want
to mount into the container. What we really want is to create two bind mounts,
one for "/usr/src/linux-hwe-6.2-headers-6.2.0-35" and another for
"/usr/src/linux-hwe-6.2-lib-rust-6.2.0-35-generic/rust".

This commit fixes this. It does so by modifying function CreateMountSpec() in
utils.go to find the first-level subdirs under "/usr/src", and then for each
such subdir, find the lcp among it's lower-level directories.

FYI: I caught this problem while running:

$ docker run --runtime=sysbox-runc -it --rm ghcr.io/nestybox/fedora:31

on my Ubuntu Jammy host in which the kernel headers for rust where in a
different directory (but still under /usr/src) than the rest of the headers.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>