Remove osusergo build tag.
Remove the "osusergo" build tag in the sysbox-mgr Makefile to work-around
a problem where sysbox-mgr was hanging during start, when doing a pre-flight
shiftfs test.

The shiftfs test was hanging when we spawned a child process and the
child process executed "unix.Setresuid(0, 0, 0)". The hang started to
occur after a "go mod tidy" on the sysbox-mgr go.mod file updated
golang from v1.18 -> v1.21.

It's not clear why the hang occurs, but it's clear that removing the
"osusergo" build tag resolves the problem.

Per the golang docs for the os/user package (https://pkg.go.dev/os/user), the
"osusergo" flag has this definition:

"For most Unix systems, this package has two internal implementations of
resolving user and group ids to names, and listing supplementary group IDs. One
is written in pure Go and parses /etc/passwd and /etc/group. The other is
cgo-based and relies on the standard C library (libc) routines such as
getpwuid_r, getgrnam_r, and getgrouplist.

When cgo is available, and the required routines are implemented in libc for a
particular platform, cgo-based (libc-backed) code is used. This can be
overridden by using osusergo build tag, which enforces the pure Go
implementation."

Unfortunately, a direct effect of removing the "osusergo" tag is that the
sysbox-mgr binary is no longer fully static, but relies on dynamic linking with
libc. This however should not a problem for sysbox-mgr, since it does not
perform low-level ops on the host system (compared to sysbox-runc or sysbox-fs).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>