Revert "Remove osusergo build tag."
Add the osusergo compilation flag in the sysbox-mgr Makefile.

We previously removed this flag to avoid a hang when sysbox-mgr calls
unix.setresuid(). By removing the flag, we forced the Go compiler to use the
Glibc wrappers for setresuid() rather than a pure Go implementation (the latter
causing the hang because it directly calls the Linux setresuid() syscall which
only changes the uid of the calling thread rather than the whole process).

However, the removal of the osusergo flag also causes sysbox-mgr to no longer be
a fully static binary, creating Glibc dependency problems that make it hard for
sysbox-mgr to run on different Linux distros.

Fortunately we can now add the osusergo flag back due to a recent change in
sysbox-libs which ensures that Sysbox always uses the Glibc wrapper for
setresuid() (and similar).

This has the benefit of making sysbox-mgr a fully static binary, removing
dependencies to run on different Linux distros.