Implement 'relaxed-read-only' mode to ease the bind-mount requirements of RO containers
Our purpose is to relax the 'read-only' requirements imposed on rootfs
mountpoints when containers are created with the `--read-only` oci-spec flag.

Prior to this change, Sysbox was blindly honoring the `--read-only` request and
exposing all the rootfs bind-mounts as RO, including all Sysbox's special mounts
and with the exception of `/proc/sys`.

The above approach prevents the execution of multiple apps within the sys
container (dockerd being one of them since it can't write to `/sys/fs/cgroup`,
`/run`, or `/tmp`).

I'm creating a global sysMgr's cli flag to allow users to decide when to enable
this behavior, which is deactivated by default.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>