Change the way Sysbox detects container removal.
The sysbox-mgr needs to detect when a container is removed to perform some
cleanup actions. Since the higher-layer runtime (e.g., Docker ) does not
explicitly notify Sysbox when a container is removed, sysbox-mgr checks for
container removal by monitoring if the container's rootfs is present or removed.

But turns out this does not work when Sysbox is doing rootfs cloning and Docker
is configured with the containerd image store. In this case, Docker's attempt to
remove the container's rootfs fails because Sysbox has rootfs clone mounts on
top of the rootfs. Thus Sysbox never detects the container removal and therefore
never cleans up the rootfs cloning mounts it set up for the container, leaving
stale mounts on the host.

As a work-around, this commit changes the way sysbox-mgr watches for container removal
when using rootfs cloning, by checking if the rootfs' overlayfs "workdir"
(e.g., workdir=/var/lib/docker/overlay2/<uuid>/work) becomes empty. This
works well, whether Docker uses the Docker image store or the containerd image
store.

The change only affects rootfs cloning. When Sysbox is not using rootfs cloning
(e.g., it's using shiftfs or idmapped mounts on overlayfs) it continues to watch
the containre's rootfs to detect container removal.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>